# cql3: select_statement: don't copy metadata object needlessly